requirements.txt
caldav==0.1.12
ldapper==0.8.3
pycli_tools==1.5
python-dateutil==1.1
python-ldap==2.4.10
vobject==0.8.1c
